requirements.txt
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1633637554808/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1634853077946/work
pandas==1.3.4
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1633452062248/work
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1632611470401/work
scikit-surprise==1.1.1
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy_1628206569514/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1633102299089/work
